uimaest-gc-24b3eeac/uimaest-gc-24b3eeac-backend/ms1/requirements.txt
uvicorn
fastapi
aider-chat
langchain
langchain_openai
langchain_groq
langchain_core
transformers
aiofiles
python-multipart